fix(aisearch): fix aisearch crashing on status error 500
