+++ src/components/docs/AiSearchHelpers/constant.js
@@ -1,0 +1,30 @@
+export const Lang = [
+  {
+    name: 'EN',
+    value: 'english',
+  },
+  {
+    name: 'ES',
+    value: 'spanish',
+  },
+  {
+    name: 'TL',
+    value: 'tagalog',
+  },
+  {
+    name: 'HI',
+    value: 'hindi',
+  },
+  {
+    name: 'IT',
+    value: 'italian',
+  },
+  {
+    name: 'FR',
+    value: 'french',
+  },
+  {
+    name: 'DE',
+    value: 'german',
+  },
+];
